Create Configuration class on Controller package.
