+++ src/main/java/com/resources/model/Story.java
@@ -2,8 +2,10 @@ package com.resources.model;
 
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
+
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
+
 import lombok.Data;
 
 @Document(collection = "stories") 
@@ -13,5 +15,4 @@ public class Story {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private String _id;
-	
 }
